requirements.txt
fastmcp
openai
python-dotenv
mcp
